+++ newton/_src/solvers/featherstone/kernels.py
@@ -348,6 +348,7 @@ def jcalc_tau(
     joint_target_kd: wp.array[float],
     joint_limit_ke: wp.array[float],
     joint_limit_kd: wp.array[float],
+    joint_damping: wp.array[float],
     joint_S_s: wp.array[wp.spatial_vector],
     joint_q: wp.array[float],
     joint_qd: wp.array[float],
@@ -405,8 +406,11 @@ def jcalc_tau(
             target_kd = joint_target_kd[j]
             target_pos = joint_target_q[target_q_start + i]
             target_vel = joint_target_qd[j]
+            damping = joint_damping[j]
 
-            drive_f = joint_force(q, qd, target_pos, target_vel, target_ke, target_kd, lower, upper, limit_ke, limit_kd)
+            drive_f = joint_force(
+                q, qd, target_pos, target_vel, target_ke, target_kd, lower, upper, limit_ke, limit_kd, damping
+            )
 
             # total torque / force on the joint
             t = -wp.dot(S_s, body_f_s) + drive_f + joint_f[j]
@@ -1108,6 +1112,7 @@ def eval_rigid_tau(
     joint_limit_upper: wp.array[float],
     joint_limit_ke: wp.array[float],
     joint_limit_kd: wp.array[float],
+    joint_damping: wp.array[float],
     joint_S_s: wp.array[wp.spatial_vector],
     body_fb_s: wp.array[wp.spatial_vector],
     body_f_ext: wp.array[wp.spatial_vector],
@@ -1149,6 +1154,7 @@ def eval_rigid_tau(
             joint_target_kd,
             joint_limit_ke,
             joint_limit_kd,
+            joint_damping,
             joint_S_s,
             joint_q,
             joint_qd,