+++ newton/_src/solvers/semi_implicit/kernels_body.py
@@ -26,6 +26,7 @@ def joint_force(
     limit_upper: float,
     limit_ke: float,
     limit_kd: float,
+    damping: float,
 ) -> float:
     """Joint force evaluation for a single degree of freedom."""
 
@@ -45,7 +46,9 @@ def joint_force(
         damping_f = -limit_kd * qd
         target_f = 0.0
 
-    return limit_f + damping_f + target_f
+    passive_f = -damping * qd
+
+    return limit_f + damping_f + target_f + passive_f
 
 
 @wp.kernel
@@ -72,6 +75,7 @@ def eval_body_joints(
     joint_limit_upper: wp.array[float],
     joint_limit_ke: wp.array[float],
     joint_limit_kd: wp.array[float],
+    joint_damping: wp.array[float],
     joint_attach_ke: float,
     joint_attach_kd: float,
     body_f: wp.array[wp.spatial_vector],
@@ -180,6 +184,7 @@ def eval_body_joints(
                 joint_limit_upper[qd_start],
                 joint_limit_ke[qd_start],
                 joint_limit_kd[qd_start],
+                joint_damping[qd_start],
             )
         )
 
@@ -217,6 +222,7 @@ def eval_body_joints(
                 joint_limit_upper[qd_start],
                 joint_limit_ke[qd_start],
                 joint_limit_kd[qd_start],
+                joint_damping[qd_start],
             )
         )
 
@@ -256,6 +262,7 @@ def eval_body_joints(
                     joint_limit_upper[qd_start + 0],
                     joint_limit_ke[qd_start + 0],
                     joint_limit_kd[qd_start + 0],
+                    joint_damping[qd_start + 0],
                 )
             )
 
@@ -280,6 +287,7 @@ def eval_body_joints(
                     joint_limit_upper[qd_start + 1],
                     joint_limit_ke[qd_start + 1],
                     joint_limit_kd[qd_start + 1],
+                    joint_damping[qd_start + 1],
                 )
             )
 
@@ -304,6 +312,7 @@ def eval_body_joints(
                     joint_limit_upper[qd_start + 2],
                     joint_limit_ke[qd_start + 2],
                     joint_limit_kd[qd_start + 2],
+                    joint_damping[qd_start + 2],
                 )
             )
 
@@ -351,6 +360,7 @@ def eval_body_joints(
                     joint_limit_upper[i_0],
                     joint_limit_ke[i_0],
                     joint_limit_kd[i_0],
+                    joint_damping[i_0],
                 )
             )
 
@@ -397,6 +407,7 @@ def eval_body_joints(
                     joint_limit_upper[i_0],
                     joint_limit_ke[i_0],
                     joint_limit_kd[i_0],
+                    joint_damping[i_0],
                 )
             )
             t_total += axis_1 * (
@@ -412,6 +423,7 @@ def eval_body_joints(
                     joint_limit_upper[i_1],
                     joint_limit_ke[i_1],
                     joint_limit_kd[i_1],
+                    joint_damping[i_1],
                 )
             )
 
@@ -423,6 +435,7 @@ def eval_body_joints(
                 0.0,
                 joint_attach_ke,
                 joint_attach_kd * angular_damping_scale,
+                0.0,
                 0.0,
                 0.0,
                 0.0,
@@ -465,6 +478,7 @@ def eval_body_joints(
                     joint_limit_upper[i_0],
                     joint_limit_ke[i_0],
                     joint_limit_kd[i_0],
+                    joint_damping[i_0],
                 )
             )
             t_total += axis_1 * (
@@ -480,6 +494,7 @@ def eval_body_joints(
                     joint_limit_upper[i_1],
                     joint_limit_ke[i_1],
                     joint_limit_kd[i_1],
+                    joint_damping[i_1],
                 )
             )
             t_total += axis_2 * (
@@ -495,6 +510,7 @@ def eval_body_joints(
                     joint_limit_upper[i_2],
                     joint_limit_ke[i_2],
                     joint_limit_kd[i_2],
+                    joint_damping[i_2],
                 )
             )
 
@@ -535,6 +551,7 @@ def eval_body_joint_forces(
                 model.joint_limit_upper,
                 model.joint_limit_ke,
                 model.joint_limit_kd,
+                model.joint_damping,
                 joint_attach_ke,
                 joint_attach_kd,
             ],